增加xorg部分软件包编译

+++ build_iso.sh
@@ -13,23 +13,26 @@ if [ "$1" == "--clean" ];then
     exit 0;
 fi
 
-# 删除已经存在的workspace目录
-if [ -d "$LIVE_BOOT" ]; then
-    echo "$LIVE_BOOT 已经目录存在，正在删除..."
-    rm -rf $LIVE_BOOT
-    echo "目录已删除"
-fi
-mkdir -p "$LIVE_BOOT"
+if [ "$1" == "--rebuild" ];then
+    # 删除已经存在的workspace目录
+    if [ -d "$LIVE_BOOT" ]; then
+        echo "$LIVE_BOOT 已经目录存在，正在删除..."
+        rm -rf $LIVE_BOOT
+        echo "目录已删除"
+    fi
+    mkdir -p "$LIVE_BOOT"
 
-# 下载debian base
-sudo debootstrap \
-    --arch=amd64 \
-    --variant=minbase \
-    stable \
-    "${DEBIAN_CHROOT}" \
-    http://mirrors.163.com/debian/
+    # 下载debian base
+    sudo debootstrap \
+        --arch=amd64 \
+        --variant=minbase \
+        stable \
+        "${DEBIAN_CHROOT}" \
+        http://mirrors.163.com/debian/
+fi
 
 echo "cutefish-live" | sudo tee "${DEBIAN_CHROOT}/etc/hostname"
+
 
 # 安装基础软件
 sudo chroot "${DEBIAN_CHROOT}" << EOF
@@ -37,7 +40,11 @@ apt-get update && \
 apt-get install -y --no-install-recommends \
     linux-image-amd64 \
     live-boot \
-    systemd-sysv
+    systemd-sysv \
+    sudo \
+    iproute2 \
+    dbus \
+    network-manager
 EOF
 
 sudo chroot "${DEBIAN_CHROOT}" << EOF
@@ -52,32 +59,32 @@ sudo chroot "${DEBIAN_CHROOT}" << EOF
     usermod -aG sudo cutefish-live
 EOF
 
-# 安装cutefish安装器
+# # 安装cutefish安装器
 mkdir ${DEBIAN_CHROOT}/package
 cp ./package/cutefish/*.deb ${DEBIAN_CHROOT}/package/
 
-# 安装所有cutefish软件
-sudo chroot "${DEBIAN_CHROOT}" << EOF
-    cd /package
-    dpkg -i *.deb
-    rm -f /var/cache/apt/archives/*
-    apt --fix-broken -d install -y
-    cd /var/cache/apt/archives/
-    dpkg -i --force-overwrite *.deb
-    apt --fix-broken install -y
-    cd /package
-    dpkg -i --force-overwrite *.deb
-    apt remove kdeconnect -y
-    apt remove zutty -y 
-    apt remove plasma-discover -y
-    apt remove systemsettings -y
-    apt remove plasma-systemmonitor -y
-    apt remove partitionmanager -y
-    apt remove kwalletmanager -y
-    apt remove plasma-workspace -y
-    rm -f /var/cache/apt/archives/*
-EOF
-rm -rf ${DEBIAN_CHROOT}/package
+# # 安装所有cutefish软件
+# sudo chroot "${DEBIAN_CHROOT}" << EOF
+#     cd /package
+#     dpkg -i *.deb
+#     rm -f /var/cache/apt/archives/*
+#     apt --fix-broken -d install -y
+#     cd /var/cache/apt/archives/
+#     dpkg -i --force-overwrite *.deb
+#     apt --fix-broken install -y
+#     cd /package
+#     dpkg -i --force-overwrite *.deb
+#     apt remove kdeconnect -y
+#     apt remove zutty -y 
+#     apt remove plasma-discover -y
+#     apt remove systemsettings -y
+#     apt remove plasma-systemmonitor -y
+#     apt remove partitionmanager -y
+#     apt remove kwalletmanager -y
+#     apt remove plasma-workspace -y
+#     rm -f /var/cache/apt/archives/*
+# EOF
+# rm -rf ${DEBIAN_CHROOT}/package
 
 
 # 创建构建iso所需目录
